Always return checked out resources
Previously, whilst checking out resources, nothing was returned if
the checkOutResources was called a second time, checking out a
new resource.

The added behavior is that the plugin will always return the resources
being checked out. It is worth to note that this plugin will NOT set
env.RESERVED_RESOURCES. RESERVED_RESOURCES is set in BFF, using the
return value of checkOutResources.

Change-Id: I8faa85635f047d4627927956ee338ee9a810d31f
Solves: Jira RNDTS-2299